refactor: Begin changing project structure
