+++ core/include/NovelRT/Core/Atom.hpp
